Crate wolfram_alpha

A library providing Rust bindings for the Wolfram|Alpha web API.

The library provides a WolframAlphaRequestSender trait which can be implemented by various request senders. These implementations may then be used to execute requests to the API.

Response bodies are deserialized from XML into structs via the serde_xml library.

Modules

model
Struct and enum definitions of values in the Wolfram|Alpha model.

query
Performs a query to Wolfram|Alpha.

Enums

Error
Represents errors which occur while using the Wolfram|Alpha API.

HttpRequestError
Represents errors which occur when sending an HTTP request to Wolfram|Alpha.

Traits

WolframAlphaRequestSender
Functionality for sending requests to Wolfram|Alpha via HTTP.

Type Definitions

Result
A convenient alias type for results for wolfram_alpha.