Module wolfram_alpha::model

Struct and enum definitions of values in the Wolfram|Alpha model.

For more information, see Wolfram|Alpha's API documentation.

Structs

Assumption
A single assumption, typically about the meaning of a word or phrase.

Assumptions
A series of Assumption elements.

Definition
An element containing a definition of a concept used in the query result.

Definitions
A series of Definition attributes.

Delimeters
Represents a warning regarding mismatched delimiters in a query.

DidYouMean
Provides a suggestion for a different query than the one provided.

DidYouMeans
A series of DidYouMean elements.

Error
Error elements provide a code and a description of an error which has occured while performing a query to the API.

ExamplePage
Provides a link to an HTML page of sample queries on the requested topic.

FutureTopic
Provides information for queries regarding a topic under development.

Generalization
Provides a link to a generalization of the requested query.

ImageMap
An HTML image map.

Img
An HTML img element suitable for direct inclusion in a web page.

Info
A container for extra information regarding the contents of a pod.

Infos
A series of Info elements.

LanguageMsg
Generated when a foreign language is detected in a query.

Link
A standard link of some text pointing to a URL.

Notes
A series of Note attributes.

Pod
Pod elements are subelements of QueryResult. Each contains the results for a single pod.

QueryResult
QueryResult is the outer wrapper for all results from the query function.

Rect
A rectangular section of an HTML image map.

Reinterpret
Represents a warning that the query was reinterpred.

RelatedExample
Provides a link to an HTML page of related examples on the requested topic.

RelatedExamples
A series of RelatedExample elements.

Sound
Provides a link to a playable sound file.

Sounds
A series of Sound elements.

Source
A link to a web page of source information.

Sources
A series of Source elements.

Spellcheck
Provides word and suggestion attributes as alternative spellings for a word in the query.

State
An alternative state which is available to a Pod or Subpod.

Statelist
A series of State attributes, combined into a list.

States
A series of State attributes.

Subpod
A subelement of Pod.

Tip
A single tip, typically regarding suggestions for fixing the query.

Tips
A series of Tip elements.

Translation
Represents a word or a phrase which was translated in the query.

Unit
A table of unit abbreviations used inside a Pod.

Units
A series of Unit elements.

Value
The value of an assumption.

Warnings
A series of Warning elements.

Type Definitions

Alternative
An alternative interpretation of an element in a query.

Note
A note made by Wolfram|Alpha regarding a query.

Plaintext
A textual representation of a single subpod.