Trait wolfram_alpha::WolframAlphaRequestSender

pub trait WolframAlphaRequestSender {
    fn send<'a>(&self,
                method: &str,
                params: &mut HashMap<&str, &'a str>)
                -> HttpRequestResult<String>;

    fn send_authed<'a>(&self,
                       method: &str,
                       app_id: &'a str,
                       params: &mut HashMap<&str, &'a str>)
                       -> HttpRequestResult<String> { ... }
}

Functionality for sending requests to Wolfram|Alpha via HTTP.

Should be implemented for clients to send requests to Wolfram|Alpha.

Required Methods

fn send<'a>(&self,
method: &str,
params: &mut HashMap<&str, &'a str>)
-> HttpRequestResult<String>

Performs an API call to Wolfram|Alpha.

Takes a map of parameters which get appended to the request as query parameters. Returns the response body string.

Provided Methods

fn send_authed<'a>(&self,
method: &str,
app_id: &'a str,
params: &mut HashMap<&str, &'a str>)
-> HttpRequestResult<String>

Make an API call to Wolfram|Alpha that contains the configured App ID.

Takes a map of parameters which get appended to the request as query parameters. Returns the response body string.